Remove voice vote muting
There are a couple reasons for removing this:

(1) Less code for us to maintain, as always.
(2) Adding to the above, I don't think any servers use this feature.
(3) It doesn't make sense to let normal members vote on a moderation
action, that should only be performed by mods. I wouldn't add a votekick
or voteban command, for example. Unless you wanted to play Among Us in
Discord.