Use client.fetch_user() instead of banning in id_lookup()
Discord added a proper API call that is bot-only to look up any user ID
on Discord, instead of having to do the hacky workaround of banning a
user ID and viewing the banned user. So use that API call instead, in
the path that is taken if the bot doesn't see the given user ID on
Discord.

Also, type hints have been added to id_lookup().