Add logs.py
Whoops, since this was a new file I forgot to add it in the last commit.
I use `git commit -a` and apparently that doesn't add untracked files,
so now I know.

What's worse is that I deleted log_added_uncached_reaction() and
log_removed_uncached_reaction() from the file before I staged it, and
when I realized that it wasn't staged, I tried to undo the removals, but
my history was gone. So that was a bit bad.