Update Pillow version check in \version
Kinda ironic whenever methods to check library versions change between
versions, but PIL.VERSION was removed from Pillow 6.0.0. That specific
constant was useless anyway, it was always the last old PIL version for
compatibility and never the actual Pillow version, we could've used
PILLOW_VERSION instead for that. But now it's __version__ (which also
works before 6.0.0).