Remove \math command
This is removed for reasons similar to
c25eb63aa71148d4f444a281e313e3c82b3418f9:

(A) It's not useful when everyone can just open a calculator on their
    own device, and it's very rudimentary - it doesn't support a lot of
    operations, requires spaces around operators, people sometimes try
    stuff like 1+1 and get a vague message, then try to find out if
    there's any support for π, e, !, sqrt, formulas with more than 2
    terms with brackets, complex numbers...
(B) Less code for us to maintain. Better focus on the things our bot is
    actually good and special at.
(C) There's an embarrassingly obvious way to DoS the entire bot with it.

I also want to mention the new slash commands supporting only 50 or 100
commands didn't influence this. We're not planning to replace our bot
commands system with slash commands.