Remove hangman
The refactored hangman that was meant to be used in more than one
hardcoded never got real use. Here's the reasons this is getting
removed:

(A) It was almost finished, except for a VERY MINOR detail that made
    games just a LITTLE unbalanced. I'll just quote someone who was
    setting up a game: "okay everyone close your eyes?"
    And it's not as simple as you might think - you know that bar at
    the top of [\]'s response to commands? Don't look at that either.
(B) It was kinda spammy - maybe for today's standards? This could work
    better with reactions anyway.
(C) Again, less code to maintain. Especially true because the
    on-message event handler no longer needs to check for a "hangman
    invoker".