Relicense under AGPLv3-only, add/use SPDX headers
This bot is now licensed under the Affero General Public License. This
is because the normal GPL only requires distribution of modified source
code if someone were to run the program on their computer, which is not
the same thing as interacting with someone's program over a network.
Given that the point of [\] is to be a Discord bot, this is a gaping
loophole that goes against the spirit of the GPL. Thus, the AGPL steps
in to patch up this exploit.

And like the previous license, this will be AGPLv3-only, rather than
AGPLv3-or-later, because it's a bad idea to include the entirety of the
Free Software Foundation in the attack surface of the license.

Additionally, every source code file now has headers denoting
AGPLv3-only, just to make sure. We've gotten kind of lazy about
including those in each new source file we've created. Furthermore, they
are SPDX headers, which is better than the boilerplate headers the FSF
recommends using.

Additionally, license.md has been renamed to LICENSE.md, and readme.md
has been renamed to README.md. There is also a LICENSE.spdx file.