Remove nitrobooster config value in favor of discord.py 1.6 features
discord.py 1.6 adds discord.Guild.premium_subscriber_role, to grab the
Nitro Booster role from a guild, and also
discord.Role.is_premium_subscriber() to determine if a role is the Nitro
Booster role or not. This thus removes the need for heuristics to detect
the booster role and to keep it in a config value.