Use Guild.fetch_ban() instead of iterating through Guild.bans()
This saves time and memory usage whenever a user gets banned and the
guild has a ban logs channel enabled. Obviously, it still requires the
Ban Members permission. Imagine doing Guild.bans() in a guild with
hundreds or even thousands of bans!

Guild.fetch_ban() can return NotFound, so in that case I decided to add
another fallback message that said the user could not be found. I also
decided to handle HTTPException, same as Forbidden, just in case.