Call reaction_roles.db_load() when reloading bot
It turns out that not doing this results in reaction_roles.connection
and reaction_roles.cursor being None after a reload.

The things that need to be reloaded should probably be de-duplicated
somewhere, because it's not obvious that this is something that you need
to take care of whenever you add a new module with its own database. But
that can be fixed later.