Remove "colors xd" and export PYTHONPATH comments from intro.sh
We don't need these comments. Also I have no idea what the PYTHONPATH
thing is about, but we have version control, we're not primitive
creatures that need to comment out a line of code if we think we need it
but want to keep it around in the future just in case.