Defer sys.exit() until client.run() returns
The problem with \kill and \restart was that it would call sys.exit()
while the client loop would be attempting to clean everything up, and it
would get real messy. Then it would exit with code 1, instead of code 4
or 8 as intended.

To fix this, we would have to move the sys.exit() call until after the
client.run() in main.py. Unfortunately, there seems to be no good way to
pass the exit code all the way from \kill and \restart back to main.py,
so I'll just have to add an exit_code variable to main.py, and use
__main__.exit_code in \kill and \restart. It works just fine.