Change reaction roles permission check to be guild-wide
Apparently, you shouldn't use a specific channel to check for the
guild-wide Manage Role permission, because Manage Roles is the exact
same permission as Manage Permissions for a channel (being able to
manage channel overrides and whatnot), despite the two functionalities
being somewhat distinct, with one being guild-wide and the other being
channel-specific.

So now [\] will correctly identify that it has guild-wide Manage Roles
if the channel it has reaction roles in disables its Manage Permissions.