Allow passing reason through givetakeroles()
I'm working on being able to limit the amount of reaction roles a member
can have from a given group of reaction roles at a time, and to do that,
the bot will have to add and remove roles at the same time if the member
hits the limit.

The best function for doing this is givetakeroles(), but it doesn't
support passing a reason through. So now it supports passing a reason
through.