Fix missing comma in reaction_roles.remove_message()
If there isn't a comma here, then it means you are treating the string
like a function and attempting to call it, which won't work. Python 3.8
has added a SyntaxWarning for this, and I'm running Python 3.9 and
noticed it, so now it's fixed.

The function is unused, and therefore untested, so of course errors like
this will slip through.