Add max_count parameter to reaction roles and overhaul cooldown system
Coral requested to be able to limit the amount of roles a member could
get from a given set of reaction roles at a time. So I've implemented
that.

The trickier part was overhauling the cooldown system to work with it,
because right now it only has a cooldown for removing a reaction role,
and the max_count thing could potentially remove a role at the same time
it adds one (because the member could exceed max_count by reacting to a
message).

So I present to you, the new cooldown system: any initial action taken
by a member will apply immediately (even if it's an unreact), then there
will be a 2-second cooldown window. During this period, any actions
taken by the member will be collected and placed into a queue, which
will take effect 1 second after the cooldown is complete.

This means we prevent role change spamminess, but also don't
unnecessarily delay adding or removing a role from a member unless we
have to (unlike the previous system, which ALWAYS delayed removing a
role from a member).

Lastly, if the member exceeds the new max_count parameter, we unreact a
few seconds after we modify their roles; otherwise we'll pick up our own
unreact event and trip over ourselves. (Discord has no way to
differentiate between a reaction event that you caused and a reaction
event that someone else caused; a bit of a shame, really.)