Add \emotes command
Shounic wanted a list of emotes on his guild to put into a strawpoll.
This is one of those things where Discord does not have any user
interface at all, so I've added this command to let people be able to
easily copy the list of emotes from the guild. In the future Shounic can
use this instead of asking me.