Refactor reaction roles to allow groups over multiple messages
What if you want to have a limit on the number of roles, but you've hit
the reaction limit on a single group of messages? Well, that's what this
refactor is for. Reaction role groups are no longer tied to message IDs;
instead, a group ID is stored in the database for each group, keeping
track of each message ID as well (but the group ID is the ID of the
first message just to keep things simple).

This also adds an interface to `\reactroles create` that lets you create
reaction role groups that span across multiple messages.