Add BlackHoleChannel as stand-in for deleted specialchannels
So, we got an exception due to trying to send a message to the mod log
of a server which had been deleted. We weren't really that disciplined
programmers when we wrote most of this code, so we assumed that once
configured, channels weren't going to disappear.

The BlackHoleChannel is now returned as a dummy class which has a
.send() method which just does nothing except print a message to the
console. Now this stopgap measure should have made a lot of our code
error-free until we improve it overall.