Add `\temperature` command
Can also be used as \temp.

This converts between Celsius and Fahrenheit. And before you ask for
Kelvin, I doubt anyone uses Kelvin in practical everyday conversation so
I'm not doing it. Same for Rankine which is even worse, plus all other
units of temperature that pretty much no one uses.

Adding this because a TF2 YouTuber wanted it.