Remove usage of discord.Embed.Empty
This solves the "Embed Changes" section. It's hard to check whether
we're not comparing two embeds or using them as dictionary keys, but
that seems a bit absurd so if everything works after testing I'll say
this is solved. (If we do happen to get crashes later, we can always
see it's about an Embed object and look at the guide again)