Workaround 1024 chars limit in starboard file list
Today I saw a message fail to get on the starboard because:

  discord.errors.HTTPException: 400 Bad Request (error code: 50035):
  Invalid Form Body
  In embed.fields.0.value: Must be 1024 or fewer in length.

The reason was that there were 5 attachment links, which apparently
broke the char limit. Removing the links and only using the filenames
made the message post successfully.

I'm guessing it's because of the monstrous extra security parameters
that Discord adds to attachment URLs nowadays, so a better solution
would be to strip those off, but this is at least better than nothing
for now...

Gosh that 1024 chars limit has silently broken so many webhook messages
of mine too, it's a pain