Stop message edit events if edit time is too old
Recently, Discord has been spamming our moderation logs with, to us,
useless message edit events. The gist is that these events are being
sent out whenever teens are looking through old messages with images in
them, causing Discord to scan those images for explicit content, and
notify clients that the content_scan_version has been changed. We can't
simply check if that's the only property that has changed, since all of
these updates take place on uncached messages, so we only get the most
recent version of the message.

The solution here is that the editing timestamp will not be changed by
this content scanning update, so it will be set to never or far in the
past.

I've also fixed `AttributeError: 'NoneType' object has no attribute
'id'` on the line `author_id = author.id` - this was funnily enough
seemingly impossible because it means a message was edited by someone
who is not in the server, but naturally this can happen with the new
content scanning updates.