Use Message.is_system() instead of MessageTypes
We used to determine that a message was a system message by its type
being different than MessageType.default. Since discord.py 2.0, there's
now the additional type MessageType.reply, which is otherwise the same
as a default message.

Instead of checking for both types, it turns out there's a
Message.is_system() function, so I'll just switch to that instead.

This should cover the "Miscellaneous Changes" section of the migration
guide.