Remove \bans
This needs code changes to work in discord.py 2.x, I looked at it more
critically and it breaks if the server has too many bans, because the
response message gets too long. The reason the command was added in
2016 was for transparency (anyone can run it, not just moderators) but
this may not be expected behavior to be left visible and leak
information that's otherwise locked behind permissions by default. We
made this decision for our server because the bot was only for our
server to use, but now it's a bot that other servers can use too.
People who _do_ have access to the ban list already can just look in
server settings. Due to these reasons, it's probably better to remove
it.