Make datetime.now() in raw message edit event UTC
This fixes `TypeError: can't compare offset-naive and offset-aware
datetimes`.

So... This timezone-naive and timezone-aware stuff is a bit confusing,
I think what's happening is when I convert the ISO timestamp from the
raw API payload (which contains a timezone) that's an offset-aware
timestamp, and datetime.datetime.now() (without argument) is
offset-naive. And if you fill in an explicit timezone, _then_ it's an
offset-aware timestamp, even though it has all the information to make
an aware timestamp, it doesn't...

This commit should only be necessary in 1.7-backports because 2.x has
discord.utils.utcnow().