Don't use Message/Member/etc objects after edit
This resolves the "Removing In-Place Edits" section in the discord.py
2.x migration guide.

The gist of this change is:

Before:
	await member.edit(nick='new nick')
	await member.send(f'Your new nick is {member.nick}')

After:
	updated_member = await member.edit(nick='new nick')
	await member.send(f'Your new nick is {updated_member.nick}')

In other words, after an edit, the old object isn't updated anymore to
reflect the change, and instead the edit method returns an updated
object.

However, this leaves me with a couple different possible
interpretations, since this is not documented too well in my opinion?
First off, why wouldn't you just always do this:

	member = await member.edit(...)

This would avoid having to use different names like member,
updated_member, updated_member2, etc. I would *guess* obviously you can
do that and it has no adverse effects (unless you wanted to use values
from the old object for reference after editing it). But I'm just not
nerdy and up-to-speed enough in python and discord.py to consciously
realize what kind of by-value/by-reference/cache implications it could
possibly have, since the documentation does this weird roundabout way
with keeping around two objects and avoids documenting overwriting the
object, it seems like you're *supposed* to do it that way? Oh well, I'm
probably overly paranoid with this, but the only thing we do with
objects after editing them is grabbing their IDs and mentions, so we
could as well do that before the edit and then don't touch it again, to
be completely safe.

Another problem with this returning-new-object documentation is that it
sometimes only returns an object "if applicable", based on what
arguments you give it, and then it's not fully documented what all
those arguments are. So for Member.edit, passing `suppress` means you
get a None back, but there are also other possible arguments? So you
may try to do updated_member.id and then oops. Anyway, we might as well
only use either the old object or the return value if we need to (like
sending a member a DM after editing them or something).