Account for abc.GuildChannel.overwrites Objects
This should be the only thing to do for "Attribute Type Changes".

abc.GuildChannel.overwrites keys could be Role or Member, and can now
also be Object in case of cache lookup failure. These do not have a
`name` attribute, so I added a check for that in the place where we
used it.

I also removed a line that appended a discriminator, since it's pretty
useless nowadays. You might argue that bots still have discriminators,
but it's such an edge case that the bot username is not enough and you
need to know what specific variant of the name it is (in particular,
people abusing their bot username to pretend to be another bot so that
our bot displays an ambiguous message somewhere?) This is probably not
something world-ending.