Remove \info command
This, misleadingly, lists permissions for the specified user in the
current channel, and does nothing else. Like \bans - for server admins,
it's probably better to just check the server settings to see the
permissions than to fill the chat with an entire screen of permissions,
and when used by non-admins, it might come as a surprise that they can
call this command and reveal permissions setups (filling the chat with
an entire screen of permissions).

I know permissions could technically be obtained by anyone via the API,
but I don't really see this mostly forgotten command helping more than
it can hurt. Also, we need to keep updating strings with new
permissions - we're currently already missing several.

This also makes \info an alias of \findu.

Fixes #13.