Rewrite utils.match_input
This started as an attempt to adapt to discriminators no longer
mattering, and evolved into rewriting the entire matching system, since
I found myself trying to improve the matching priority with usernames
now being much easier to conclusively type in, and with global
nicknames also being added to the mix.

The priority is now:
1) ID or mention
2) Exact name, case-sensitive
	For members, this is their unique [a-z0-9_.] username.
	For other objects, this can be anything, and is not necessarily
	unique.
3) The following algorithm:
	for (full, startswith, contains):
		for (cs, ci):
			for (name, server_nick, global_nick):
				return if this matches

The comment goes into more detail.

The main difference from the old system is that the priority is now
simpler, certain steps aren't skipped (such as case-sensitive nickname
checking before case-insensitive), and before checking for partial
matches, there's now a check whether the name starts with the search
term (so that "ragon" will match "Ragonfara" before "Endr Dragon").