Add safety hasattr in on_member_remove
We're assuming AuditLogEntry.target always has an id attribute.
Its type is "Any" ("depends on the action being done"), so let's be
sure to check if it exists before we get problems here from random
audit log events just existing while a member is removed from the
server (causing the modlog message to not be posted).