Remove "Syncing invite cache" print
It's printing every 30 seconds, so it has the potential to bury useful
messages.

Actually, this makes me realize it hasn't been printing for a while on
production [\]... So there's probably something wrong in case of
reconnects or something. Still, now that we know that, the print should
probably be removed anyway.