Enable discord.py's root_logger
So that was what that logging change meant...

Without changes, anything logged by discord.py itself would appear
twice - one time colored and with a date, and one time in a more basic
form. This is because we were already setting up logging ourselves, via
logging.basicConfig(). It took a while to get it to show all messages
once and to also show them at all, because let's ignore how I was
trying to pass Client.run() kwargs to the discord.Client() constructor?
But now everything uses the colored discord.py logger.