Make utils.id_summary() usages more consistent
They used to sometimes be in an embed field, and sometimes in the
footer. I don't remember particularly why, but I think we tried to put
it in an embed field as often as possible and only fell back on the
footer when it "wasn't possible", for some reason. We probably wanted
to always show a timestamp in the embed, and always have it below the
ID summary, which is impossible without putting the ID summary in an
embed field.

But, putting the ID summary in the footer is better, and we're probably
not gonna keep the dates in the embeds anymore (they're shown on each
message on hover anyway, and the cleaner footer is probably worth it)