Remove timestamps from log messages
One of the things our log messages have been inconsistent at is whether
they display a timestamp at the bottom or not. The timestamps are not
*really* necessary, at least less so than they used to be, since on
desktop you can hover over any message to get the time, and even on
mobile it's not hard to get it if you really need to know the specific
time of a log message that was below a recent enough message that
Discord grouped it below that. I asked Misa, and she didn't think dates
were needed.

The reason for not having dates is that it makes the messages cleaner,
especially now that the IDs are always consistently in the footer. The
timestamp would always be to the right of the IDs, and probably
wordwrap somewhere inside the date and time in an bad-looking way
inconsistently between different types of log messages. And now the
messages are more consistent - instead of some messages having
timestamps and others not having them.