Remove explicit Python 3.9 from run.sh
Python 3.9 used to be an upgrade from the default Python 3.8 on my
system, but I've done an OS upgrade and now I have 3.12, so the .9
should go.

It's still probably best to keep the 3 there - if ever Python 4 comes
out in the future, it'll probably have breaking changes that will need
a look before we can switch to it anyway. It makes it clear that this
is a Python 3 project, instead of a really old python2 project. And
while plain `python` may not inherently mean `python2` anymore, it now
seems to be that plain `python` doesn't always exist and you have to
use `python3` explicitly to be sure you run any Python at all...