Fix SyntaxWarnings from invalid escape sequences
Apparently this has been a DeprecationWarning for a long time
(since Python 3.6 even, to actually become a SyntaxError later) but
DeprecationWarnings are silenced by default, so they changed it to
SyntaxWarning in 3.12 so more people would see it.

And, yeah we kind of had a lot of them lol. All of them should now be
fixed, either by just adding another backslash or by switching strings
to r'' ones.