Auto-fix role cache errors instead of warning
If, during startup, the bot notices that one or more users in a server
with the role cache enabled have different roles than was in the role
cache, the bot would send a warning into the modlog like "<user> has
different roles than in the cache! Maybe you want to correct things."

The solution here is to run \rolesync in the server, and that's it.
But an admin of the server has to know that (and that knowledge has to
not have faded away after years since seeing that message).

Presumably we weren't sure the current roleset was always better than
the saved one for some reason, and we as bot developers were gonna be
the only ones seeing the message anyway... But it'd be better to just
autofix it and not bother server admins with a confusing message that
"something is broken", when they don't have to be.