Add timeout modlog event
The moderation log now logs if a member has received a timeout, or if
the timeout has been (manually) revoked. Discord doesn't send an event
when a timeout expires, so possible future work is to set an expiry
timer for ourselves and check if the timeout indeed just expired.

This adds the logid member_timeout. I didn't even have to one-off add
it to any servers, because I was going to duplicate member_ban but all
servers were using member_* instead.